+++ changelog.md
@@ -7,7 +7,7 @@
 #### Modification interne
 
 - up.php :  cleanup checkfiles
-- install.php : cleanup checkfiles
+- install.php : cleanup checkfiles, clean up updated actions before installing new ones
 
 ## 06/11/2025 - version 5.4.1